remove reference to private repository
NTIA/sea-data-product is private, has been archived, and was already stale (i.e., not a useful reference to include here). Making this change ahead of potential publication or wider sharing of this repository. If a replacement reference is desired, I recommend adding a reference to the first SEA test plan